requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
fastai==2.7.14
torch==2.2.1
torchvision==0.17.1
numpy==1.24.3
Pillow==10.3.0
python-multipart==0.0.9
starlette==0.37.2
uvicorn[standard]==0.30.1
python-dotenv==1.0.1